requirements.txt
ultralytics>=8.0.0
torch>=2.0.0+cu118
torchvision>=0.15.0+cu118
torchaudio>=2.0.0+cu118
flask>=2.0.0
flask-cors>=3.0.10
opencv-python>=4.5.0
pillow>=8.0.0
numpy>=1.21.0
pyyaml>=5.4.0
requests>=2.25.0
gunicorn>=20.1.0
matplotlib>=3.5.0
seaborn>=0.11.0
pandas>=1.3.0 